fix(autoware_behavior_velocity_planner_common): fix bugprone-narrowing-conversions warnings (#940)

+++ planning/behavior_velocity_planner/autoware_behavior_velocity_planner_common/src/utilization/util.cpp
@@ -279,7 +279,7 @@ void insertVelocity(
     std::min(static_cast<int>(insert_index + 1), static_cast<int>(path.points.size() - 1));
   for (int i = min_idx; i <= max_idx; i++) {
     if (calc_distance2d(path.points.at(static_cast<size_t>(i)), path_point) < min_distance) {
-      path.points.at(i).point.longitudinal_velocity_mps = v;
+      path.points.at(i).point.longitudinal_velocity_mps = static_cast<float>(v);
       already_has_path_point = true;
       insert_index = static_cast<size_t>(i);
       // set velocity from is going to insert min velocity later
@@ -288,7 +288,7 @@ void insertVelocity(
   }
   //! insert velocity point only if there is no close point on path
   if (!already_has_path_point) {
-    path.points.insert(path.points.begin() + insert_index, path_point);
+    path.points.insert(path.points.begin() + static_cast<std::ptrdiff_t>(insert_index), path_point);
   }
   // set zero velocity from insert index
   setVelocityFromIndex(insert_index, v, &path);
